fix(cli): redact credentials in the SQLite missing-file error
    owlsql generate --url "jdbc:postgresql://user:S3cret@host/db" --out x.ts
    # Error: SQLite database file not found: "jdbc:postgresql://user:S3cret@host/db".

Two malformed shapes reach this line with credentials intact. A second
scheme token (jdbc:postgresql://) and a missing colon (postgres//user:...)
match neither SCHEME_PATTERN nor MISTYPED_URL_CREDENTIALS_PATTERN, which
expects the credentials right after the first `scheme:/`, so detectDialect
returns sqlite for both and the password lands in stderr - terminal
scrollback and CI logs included.

Redacting at the echo rather than adding a third detection pattern: sqlite
is where every unrecognized string ends up, so this closes the next
blocklist gap as well. redactCredentials already handled both strings
correctly; it just was not called here. It moves to its own module so a
dialect can call it without importing generate.ts, which imports the
dialects; generate.ts re-exports it, so its public name is unchanged.

Fixes #278
