+++ src/cli/generate.ts
@@ -5,6 +5,7 @@ import { introspectPostgres } from './dialects/postgres.js';
 import { introspectMysql } from './dialects/mysql.js';
 import { introspectSqlite } from './dialects/sqlite.js';
 import { introspectMssql } from './dialects/mssql.js';
+import { redactCredentials } from './redact.js';
 
 export interface GenerateOptions {
   url: string;
@@ -37,29 +38,7 @@ const MISTYPED_URL_CREDENTIALS_PATTERN = /^(?![a-z]:[/\\])[a-z][a-z0-9+.-]*:\/{1
 const ADO_CREDENTIALS_PATTERN =
   /(^|;)\s*(uid|user id|pwd|password|database|initial catalog|trusted_connection|integrated security|driver|dsn)\s*=/i;
 
-// URL userinfo ("//user:pass@" or a mistyped single-slash ":/user:pass@"). The
-// single-slash form requires a multi-character scheme before the ":/" so that
-// Windows drive-letter paths ("C:/app@prod.db") are not treated as URL userinfo.
-//
-// The userinfo run is greedy up to the *last* "@" of the authority, because
-// that is the one a URL parser splits on: an unencoded "@" inside a password
-// ("root:p@ss@host") is accepted by `new URL` and by the drivers, and a
-// non-greedy match stopped at the first one and printed the rest of the
-// password (issue #251). "/", "?" and "#" end the authority, so a later "@"
-// in a path or query is never mistaken for userinfo.
-const URL_CREDENTIALS_PATTERN = /(\/\/|(?<=[a-z][a-z0-9+.-]):\/)[^/?#]*@/i;
-
-// The password value of an ADO / DSN "Pwd="/"Password=" pair. The value may be
-// wrapped in double quotes, braces, or single quotes (so an embedded ";" does
-// not terminate it); an unquoted value runs to the next ";" delimiter.
-const DSN_PASSWORD_PATTERN =
-  /((?:^|;)\s*(?:pwd|password)\s*=)("[^"]*"|\{[^}]*\}|'[^']*'|[^;]*)/gi;
-
-export function redactCredentials(url: string): string {
-  return url
-    .replace(URL_CREDENTIALS_PATTERN, (_match, separator: string) => `${separator}***@`)
-    .replace(DSN_PASSWORD_PATTERN, '$1***');
-}
+export { redactCredentials } from './redact.js';
 
 function unrecognizedUrlError(url: string): Error {
   return new Error(
